updated followage service/controller

+++ src/main/java/tn/esprit/spring/services/IFollowageService.java
@@ -3,6 +3,6 @@ package tn.esprit.spring.services;
 import tn.esprit.spring.entities.Followage;
 
 public interface IFollowageService {
-	void addFollow(int idUser, Followage followage);
+	void addFollow(long idUser, Followage followage);
 	void unfollow(Followage followage);
 }
